Make e method available in view, capitalized E method for editable chunk, accept content_key alternative for identifier
